+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -5,8 +5,8 @@
 - Implement `serde::Serialize` and `serde::Deserialize` for `Duration` using `FromStr` and `Display` implementations, when feature `derive-serde` is enabled.
 
 ### Breaking changes
-- Adds new error type `ToJsonError`, which is returned when deserializing to schema type fails.
-- Adds member `JsonError::TraceError` to `JsonError`, which adds trace information to the error produced when serializing from schema type fails.
+- Add a new error type `ToJsonError`, which is returned when deserializing a schema type fails.
+- Add the member `JsonError::TraceError` to `JsonError`, which has trace information for the error produced when serializing a schema type fails.
 
 ## concordium-contracts-common 6.0.0 (2023-05-08)